Port a bunch of changes from the Digitanks Linux port over to here. Includes 64 bit updates, a new file open/save dialog, and a few other tweaks and fixes.

+++ tengine/renderer/renderer.cpp
@@ -1356,7 +1356,7 @@ size_t CRenderer::GetTextureHeight(size_t iTexture)
 	return iHeight;
 }
 
-void CRenderer::UnloadTextureData(size_t iTexture)
+void CRenderer::UnloadTextureData(unsigned int iTexture)
 {
 	ilBindImage(0);
 	ilDeleteImages(1, &iTexture);
